requirements.txt
gradio[oauth]
requests
pandas
langgraph
langchain
langchain-core
langchain-openai
langsmith
langfuse
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-requests
python-dotenv
tenacity